requirements.txt
dash
pandas
mysql-connector-python
